combine bridge-related errors
