--- packages/web/utils/error-types.ts
@@ -1,8 +1,0 @@
-export enum ErrorTypes {
-  UnexpectedError = "UnexpectedError",
-  CreateApprovalTxError = "ApprovalTxError",
-  CreateCosmosTxError = "CreateCosmosTxError",
-  CreateEVMTxError = "CreateEVMTxError",
-  NoQuotesError = "NoQuotesError",
-  UnsupportedQuoteError = "UnsupportedQuoteError",
-}
